"Add viewpoint parameter to World and Tree draw methods for panning"
This commit adds a `viewPoint` parameter to the `draw` methods of the `World` and `Tree` classes. This change allows for panning the view by modifying the `viewPoint` when calling the `draw` method. The `Tree` class's `draw` method has also been updated to adjust the position of the tree's top segment based on the `viewPoint`.