"Add height coefficient to Tree class and implement level-wise drawing with lerp"
This commit modifies the Tree class in the items/tree.js file by adding a new heightCoef parameter to the constructor. It also modifies the draw method to calculate the top point of the tree using this heightCoef and draws the tree level-wise using a helper function lerp2D. The lerp2D function is also added to the math/utils.js file.