Change Tree class to interpolate point color and remove segment drawing
The `Tree` class has been updated to interpolate the color of the points drawn for each level of the tree. The color is now determined by the level of the point, ranging from green to darker green. Additionally, the code that drew a segment between the tree trunk and top has been removed.