"Remove graphEditor.js, add stopEditor.js for stop functionality"
This diff removes the file graphEditor.js and adds a new file stopEditor.js. The GraphEditor class has been removed and its functionality has been replaced by the StopEditor class. The StopEditor class is used for handling stop functionality and is enabled and disabled in the index.html file. The display and enable/disable methods have been added to the StopEditor class.