"Change stop creation to use lane guides instead of segments; add lane guide generation to World"
This commit changes the StopEditor class to use lane guides instead of segments when creating a new stop. It also adds lane guide generation to the World class, which generates lane guides based on the road width and roundness. The lane guides are then used in the StopEditor class to create new stops. The commit also adjusts the road width used when creating a new stop to be half the road width.