"Add marking editor classes for crossing and stop markings, update Stop and Crossing classes to use new Envelope class, and add perpendicular function to math utils"
This diff adds two new classes, CrossingEditor and StopEditor, which extend the MarkingEditor class and handle the creation and deletion of crossing and stop markings, respectively. The Stop and Crossing classes have been updated to use the new Envelope class, and the js/math/utils.js file has been updated with a new perpendicular function.