"Refactor Stop class to extend Marking class and modify Crossing class to inherit from Marking"
This commit refactors the Stop class to extend the Marking class and modifies the Crossing class to inherit from the Marking class. The Stop class constructor and draw method have been updated to reflect the changes. The Crossing class constructor has been updated to call the superclass constructor and inherit its properties.