"Add static load method to Building class and support for different marking types"
This commit adds a new static method `load` to the `Building` class which takes an object containing information about the building and returns a new `Building` instance with the provided information.

Additionally, the `Marking` class has been updated to support different types of markings, including crossing, light, marking, parking, start, stop, target, and yield. A new `load` method has been added to the `Marking` class which takes an object containing information about the marking and returns a new instance of the appropriate marking subclass based on the `type` property of the input object.